requirements.txt
asttokens==3.0.0
blessed==1.20.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
codefind==0.1.7
colorama==0.4.6
cryptography==44.0.0
executing==2.2.0
gdrive==0.1.5
google-api-core==2.24.0
google-api-python-client==2.43.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.66.0
httplib2==0.22.0
icecream==2.1.4
idna==3.10
iniconfig==2.1.0
jeepney==0.8.0
jurigged==0.6.0
oauthlib==3.2.2
ovld==0.4.6
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pypdfium2==4.30.1
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
pytest==8.3.5
pytest-qt==4.4.0
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SecretStorage==3.3.3
setuptools==59.6.0
shiboken6==6.8.1.1
six==1.17.0
uritemplate==4.1.1
urllib3==2.3.0
versioneer==0.29
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.37.1
